improve: 테스트 코드 각 클래스 별로 분리
