+++ src/test/kotlin/mine_tdd/MineTest.kt
@@ -1,13 +1,11 @@
 package mine_tdd
 
-import mine_tdd.Board.Companion.findMineCount
 import mine_tdd.cell.Cells
 import mine_tdd.cell.MineCell
 import mine_tdd.cell.NoneCell
 import mine_tdd.cell.Position
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 
 class MineTest {
     @Test
@@ -47,14 +45,6 @@ class MineTest {
     }
 
     @Test
-    fun `Position()에는 음수를 입력할 수 없다`() {
-        val x = -1
-        val y = 0
-
-        assertThrows<IllegalArgumentException> { Position(x, y) }
-    }
-
-    @Test
     fun `셀의 위치 정보를 통해 찾을 수 있다`() {
         val position = Position(0, 0)
 
@@ -83,20 +73,6 @@ class MineTest {
         val cells = board.getColumnCells(column)
 
         Assertions.assertThat(cells.size()).isEqualTo(height)
-    }
-
-    @Test
-    fun `Positiion(1,1) 주변 Position의 지뢰 개수를 구할 수 있다`() {
-        val position = Position(1, 1)
-        val mineCellList = listOf(
-            MineCell(Position(0, 0)),
-            MineCell(Position(1, 0)),
-            MineCell(Position(4, 0)),
-        )
-
-        val count = position.findMineCount(mineCellList)
-
-        Assertions.assertThat(count).isEqualTo(2)
     }
 
     @Test
